// src/app/swagger/model/printingDevice.ts
/**
 * PrintingTracker REST API
 * Това REST API дефинира интерфейс за менажиране на база данни от клиенти, принтерите с които те разполагат, и броячи на използваните консумативи.  Този проект е част от дипломна защита на Мария Денкова и Иво Пеев в Технически Университет - София, катедра Компютърни Системи, випуск 2018. 
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface PrintingDevice {
    serialNumber: string;
    brand: string;
    model: string;
    dateInstalled?: Date;
    addressId: number;
    customerUic?: string;
    pricePackageId?: number;
}